rename the module to extract_skip_diff
