+++ crates/macro/src/extract_skip_diff.rs
